Split oversized header-actions and command-palette components
