+++ src/components/command-palette.ts
@@ -15,7 +15,6 @@ import { customElement, query, state } from "lit/decorators.js";
 import type { ESPHomeAPI } from "../api/index.js";
 import type { ConfiguredDevice } from "../api/types/devices.js";
 import type { LanguageChoice, LocalizeFunc } from "../common/localize.js";
-import { LANGUAGES, languageLabel } from "../common/localize.js";
 import {
   apiContext,
   devicesContext,
@@ -23,14 +22,10 @@ import {
   yamlDiffButtonContext,
 } from "../context/index.js";
 import { espHomeStyles } from "../styles/shared.js";
-import { navigate } from "../util/navigation.js";
 import { registerMdiIcons } from "../util/register-icons.js";
-import {
-  forEachYamlMatch,
-  yamlEmptyMessageKey,
-  yamlHitHref,
-  yamlHitLabel,
-} from "../util/yaml-search-helpers.js";
+import { yamlEmptyMessageKey } from "../util/yaml-search-helpers.js";
+import type { CommandAction } from "./command-palette-actions.js";
+import { buildCommands, buildYamlHitActions } from "./command-palette-actions.js";
 import { commandPaletteStyles } from "./command-palette.styles.js";
 import { YamlSearchController } from "./yaml-search-controller.js";
 
@@ -57,21 +52,6 @@ function closeAllOpenPopovers(root: Document | ShadowRoot) {
   for (const el of root.querySelectorAll<HTMLElement>("*")) {
     if (el.shadowRoot) closeAllOpenPopovers(el.shadowRoot);
   }
-}
-
-interface CommandAction {
-  id: string;
-  group: string;
-  label: string;
-  /** MDI icon name registered via ``registerMdiIcons``; rendered
-   *  through ``<wa-icon library="mdi">``. Mutually exclusive with
-   *  ``flag`` — when both are set, ``flag`` wins. */
-  icon?: string;
-  /** Emoji prefix shown in place of the MDI icon column. Used for
-   *  language entries so the picker reads as flags-not-icons. */
-  flag?: string;
-  keywords?: string[];
-  run: () => void;
 }
 
 @customElement("esphome-command-palette")
@@ -179,89 +159,14 @@ export class ESPHomeCommandPalette extends LitElement {
   };
 
   private _allCommands(): CommandAction[] {
-    const t = this._localize;
-
-    const nav: CommandAction[] = [
-      {
-        id: "nav.home",
-        group: t("command_palette.group_navigation"),
-        label: t("command_palette.go_dashboard"),
-        icon: "home",
-        keywords: ["dashboard", "devices"],
-        run: () => navigate("/"),
-      },
-      {
-        id: "nav.secrets",
-        group: t("command_palette.group_navigation"),
-        label: t("layout.secrets"),
-        icon: "key-variant",
-        keywords: ["password", "wifi"],
-        run: () => navigate("/secrets"),
-      },
-    ];
-
-    const deviceGroup = t("command_palette.group_devices");
-    const devices: CommandAction[] = this._devices.map((d) => ({
-      id: `device.${d.configuration}`,
-      group: deviceGroup,
-      label: d.friendly_name || d.name || d.configuration,
-      icon: "chip",
-      keywords: [d.configuration, d.name],
-      run: () => navigate(`/device/${d.configuration}`),
-    }));
-
-    const themeGroup = t("layout.theme");
-    const themes: CommandAction[] = [
-      {
-        id: "theme.light",
-        group: themeGroup,
-        label: t("layout.theme_light"),
-        icon: "weather-sunny",
-        keywords: ["light", "theme"],
-        run: () => this._setTheme("light"),
-      },
-      {
-        id: "theme.dark",
-        group: themeGroup,
-        label: t("layout.theme_dark"),
-        icon: "weather-night",
-        keywords: ["dark", "theme"],
-        run: () => this._setTheme("dark"),
-      },
-      {
-        id: "theme.system",
-        group: themeGroup,
-        label: t("layout.theme_system"),
-        icon: "theme-light-dark",
-        keywords: ["system", "auto"],
-        run: () => this._setTheme("system"),
-      },
-    ];
-
-    const editor: CommandAction[] = [
-      {
-        id: "editor.yaml_diff_button",
-        group: t("layout.editor"),
-        label: this._yamlDiffEnabled
-          ? t("command_palette.hide_yaml_diff_button")
-          : t("command_palette.show_yaml_diff_button"),
-        icon: "vector-difference",
-        keywords: ["diff", "yaml", "compare"],
-        run: () => this._toggleDiffButton(),
-      },
-    ];
-
-    const languageGroup = t("command_palette.group_language");
-    const languages: CommandAction[] = LANGUAGES.map((l) => ({
-      id: `language.${l.value}`,
-      group: languageGroup,
-      label: languageLabel(l, t),
-      flag: l.flag,
-      keywords: ["language", "locale", l.value],
-      run: () => this._setLanguage(l.value),
-    }));
-
-    return [...nav, ...devices, ...themes, ...languages, ...editor];
+    return buildCommands({
+      t: this._localize,
+      devices: this._devices,
+      yamlDiffEnabled: this._yamlDiffEnabled,
+      setTheme: (theme) => this._setTheme(theme),
+      setLanguage: (lang) => this._setLanguage(lang),
+      toggleDiffButton: () => this._toggleDiffButton(),
+    });
   }
 
   /**
@@ -307,30 +212,8 @@ export class ESPHomeCommandPalette extends LitElement {
     });
   }
 
-  /**
-   * Materialise the live YAML-content hits as ``CommandAction``s
-   * so the existing render + keyboard-nav code handles them
-   * without a parallel branch. Each match becomes its own row
-   * (one device with three matches → three rows) so the user can
-   * pick the specific line they want to land on. Click → navigate
-   * to ``/device/<configuration>?line=<n>``; the editor's
-   * ``_readUrlLine`` already wires that param to scroll-to + the
-   * existing highlight machinery.
-   */
   private _yamlHitActions(): CommandAction[] {
-    const groupName = this._localize("command_palette.group_yaml_matches");
-    return forEachYamlMatch(this._yamlSearch.hits, (hit, match) => ({
-      id: `yaml.${hit.configuration}:${match.line_number}`,
-      group: groupName,
-      label: yamlHitLabel(hit, match),
-      icon: "code-braces",
-      // No ``keywords`` — YAML mode bypasses ``_filtered()``'s
-      // keyword search entirely (the backend already did the
-      // matching) so an unused keywords array would just retain
-      // raw YAML line text — including potentially-sensitive
-      // values — in memory for nothing.
-      run: () => navigate(yamlHitHref(hit, match)),
-    }));
+    return buildYamlHitActions(this._yamlSearch.hits, this._localize);
   }
 
   protected willUpdate(changed: Map<string, unknown>) {
